moved offline bolt state handler to an inner class.
